+++ template/include/segtree_lazy.cpp
@@ -10,21 +10,21 @@ size n
 class $class$ {
 public:
   //!slide plugin_if sum_add
-  struct Value { long x, k; };
+  struct Value { long long int x, k; };
   Value identity_value() { return { 0, 0 }; }
   Value combine(Value a, Value b) { return { a.x+b.x, a.k+b.k }; }
 
-  struct Update { long v; };
+  struct Update { long long int v; };
   Update identity_update() { return { 0 }; }
   Value apply_update(Update a, Value x) { return {x.x + a.v*x.k}; }
   Update combine_updates(Update a, Update b) { return { a.v + b.v }; }
   //!slide plugin_end_if
   //!slide plugin_if min_set
-  struct Value { long x; };
-  Value identity_value() { return { (long)1e9+1 }; }
+  struct Value { long long int x; };
+  Value identity_value() { return { (long long int)1e9+1 }; }
   Value combine(Value a, Value b) { return { min(a.x, b.x) }; }
 
-  struct Update { long v; };
+  struct Update { long long int v; };
   Update identity_update() { return { -1 }; }
   Value apply_update(Update a, Value x) {
     if (a.v == -1) return x;
@@ -36,12 +36,12 @@ public:
   }
   //!slide plugin_end_if
 
-  long n;
+  long long int n;
   vector<$class$::Value> tree;
   vector<$class$::Update> lazy;
 
   // build segtree of size at least min_n
-  $class$(long min_n)
+  $class$(long long int min_n)
     : n(next_power_of_two(min_n)),
       tree(2*n, identity_value()),
       lazy(2*n, identity_update()) {}
@@ -51,75 +51,75 @@ public:
     : n(next_power_of_two(base.size())),
       tree(2*n, identity_value()),
       lazy(2*n, identity_update()) {
-    for (long i=0; i<(long)base.size(); ++i)
+    for (long long int i=0; i<(long long int)base.size(); ++i)
       tree[n+i] = base[i];
     build(1, 0, n);
   }
 
   // combines all values in range [l, r)
-  Value query(long l, long r) {
+  Value query(long long int l, long long int r) {
     assert(0 <= l && l < r && r <= n);
     return query_(l, r, 1, 0, n);
   }
 
-  long next_power_of_two(unsigned x) { return 1<<__lg(x-1)+1; }
+  long long int next_power_of_two(unsigned x) { return 1<<__lg(x-1)+1; }
 
   // updates all values in range [l, r)
-  void update(long l, long r, Update a) {
+  void update(long long int l, long long int r, Update a) {
     assert(0 <= l && l < r && r <= n);
     return update_(a, l, r, 1, 0, n);
   }
 
 private:
   // applies the update to the current node
-  void apply(long pos, Update a) {
+  void apply(long long int pos, Update a) {
     tree[pos] = apply_update(a, tree[pos]);
     lazy[pos] = combine_updates(a, lazy[pos]);
   }
 
   // recomputes the value of position "pos", assuming lazy[pos]==identity_update()
-  void recompute(long pos) {
+  void recompute(long long int pos) {
     tree[pos] = combine(tree[2*pos], tree[2*pos+1]);
   }
 
   // pushes lazy values down the tree
-  void propagate(long pos) {
+  void propagate(long long int pos) {
     apply(2*pos, lazy[pos]);
     apply(2*pos+1, lazy[pos]);
     lazy[pos] = identity_update();
   }
 
   // build segtree assuming only leaf nodes are correct
-  void build(long pos, long l, long r){
+  void build(long long int pos, long long int l, long long int r){
     if (r - l == 1) // leaf: do nothing
       return;
-    long m = (l+r)/2;
+    long long int m = (l+r)/2;
     build(2*pos, l, m);
     build(2*pos+1, m, r);
     recompute(pos);
   }
 
-  Value query_(long ql, long qr, long pos, long l, long r) {
+  Value query_(long long int ql, long long int qr, long long int pos, long long int l, long long int r) {
     // completely contained: return value
     if (ql <= l && r <= qr) { return tree[pos]; }
     // not overlapping: return nothing
     if (qr <= l || r <= ql) { return identity_value(); }
     // otherwise: recurse
     propagate(pos);
-    long m = (l+r)/2;
+    long long int m = (l+r)/2;
     Value ans_l = query_(ql, qr, 2*pos, l, m);
     Value ans_r = query_(ql, qr, 2*pos+1, m, r);
     return combine(ans_l, ans_r);
   }
 
-  void update_(Update a, long ql, long qr, long pos, long l, long r) {
+  void update_(Update a, long long int ql, long long int qr, long long int pos, long long int l, long long int r) {
     // completely contained: update lazy
     if (ql <= l && r <= qr) { apply(pos, a); return; }
     // not overlapping: do nothing
     if (qr <= l || r <= ql) { return; }
     // otherwise: recurse
     propagate(pos);
-    long m = (l+r)/2;
+    long long int m = (l+r)/2;
     update_(a, ql, qr, 2*pos, l, m);
     update_(a, ql, qr, 2*pos+1, m, r);
     recompute(pos);
